+++ document.R
@@ -1,4 +1,3 @@
-devtools::document()
 if(file.exists("DESCRIPTION")) unlink("docs", recursive = TRUE)
 codemetar::write_codemeta()
 knitr::knit("README.Rmd")